+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 1.0.7
+
+### Patch Changes
+
+- update internal packages
+
 ## 1.0.6
 
 ### Patch Changes