+++ blueprint/interfaces/worker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-worker
 
+## 1.0.8
+
+### Patch Changes
+
+- update internal packages
+
 ## 1.0.7
 
 ### Patch Changes