+++ framework/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/common
 
+## 1.1.7
+
+### Patch Changes
+
+- More relations covered for compliance entity
+
 ## 1.1.6
 
 ### Patch Changes