+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 1.1.7
+
+### Patch Changes
+
+- More relations covered for compliance entity
+- Updated dependencies
+  - @forklaunch/validator@1.1.7
+  - @forklaunch/common@1.1.7
+  - @forklaunch/core@1.1.7
+  - @forklaunch/ws@1.1.7
+
 ## 1.1.6
 
 ### Patch Changes